+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The document follows the conventions described in [“Keep a CHANGELOG”](http:
 - `SVG.TextPath` now inherits from `SVG.Text` and can be manipulated the same way (#705)
 - `SVG.Text.textPath()` now returns all textPaths in the text element (#705)
 - renamed `SVG.Stop` constructor `at()` on `SVG.Gradient` to `stop()` (#707)
+- renamed `fill()` method on `SVG.Gradient` and `SVG.Pattern` to `url()` (#708)
 
 ### Fixed
 - fixed a bug in clipping and masking where empty nodes persists after removal -> __TODO!__